backend/requirements.txt
# --- Web Framework ---
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9   # Required for File Uploads

# --- Data Processing ---
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2           # Required for reading Excel files
pyarrow>=15.0.0           # Required for reading Parquet files

# --- AI & LLM ---
openai>=1.12.0            # For connecting to Mistral/OpenRouter
python-dotenv>=1.0.0      # For loading .env files

# --- Database ---
# (We are using built-in sqlite3, but SQLAlchemy is good for future proofing)
sqlalchemy>=2.0.0